lesson07/requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-google==10.13.1
apache-airflow-providers-postgres==5.11.0
apache-airflow-providers-common-sql>=1.3.1
psycopg2-binary>=2.8.0
requests